# openapi.yml
openapi: 3.1.0
x-stoplight:
  id: malmoc95sulph
info:
  title: main
  version: '1.0'
servers:
  - url: 'http://localhost:3000'
paths:
  /tasks:
    get:
      summary: Your GET endpoint
      tags:
        - task
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
                  x-stoplight:
                    id: esye4wsv4xev0
      operationId: get-tasks
      x-stoplight:
        id: 9q6rfd0mb8rqm
      parameters: []
    post:
      summary: Your POST endpoint
      tags:
        - task
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
      operationId: post-tasks
      x-stoplight:
        id: gg89tsl9ogv5q
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
  '/tasks/{taskId}':
    parameters:
      - schema:
          type: string
          format: uuid
        name: taskId
        in: path
        required: true
    get:
      summary: Your GET endpoint
      tags:
        - task
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Not Found
      operationId: get-tasks-taskId
      x-stoplight:
        id: d6ogw1upatj59
      parameters: []
    delete:
      summary: Your DELETE endpoint
      tags:
        - task
      responses:
        '200':
          description: OK
      operationId: delete-tasks-taskId
      x-stoplight:
        id: wheqz161eflx2
      parameters: []
components:
  schemas:
    Task:
      title: Task
      x-stoplight:
        id: ka2mtvt3r48ep
      type: object
      properties:
        id:
          type: string
          x-stoplight:
            id: x6952st3sqiw3
          format: uuid
          readOnly: true
        site_url:
          type: string
          x-stoplight:
            id: j4mv6zri6irqd
          format: uri
          maxLength: 100
        duration_day:
          type: integer
          x-stoplight:
            id: cf4tf4eq87fev
          minimum: 0
          exclusiveMinimum: true
        condition_query:
          type: string
          x-stoplight:
            id: rdr3rv3qtyyey
        created_at:
          type: string
          x-stoplight:
            id: ix5a27wonb2qq
          format: date-time
          readOnly: true
        created_by:
          type: string
          x-stoplight:
            id: 9idul0m1r4uyy
          format: uuid
          readOnly: true
tags:
  - name: task